Fix: Betty code error and add selection sort implementation

+++ 0-bubble_sort.c
@@ -9,17 +9,17 @@
 
 void bubble_sort(int *array, size_t size)
 {
-	unsigned int i, j, temp;
+	size_t i, j, temp;
 
 	for (j = 0; j < size - 1; j++)
 	{
 		for (i = 0; i < size - 1; i++)
 		{
-			if (array[i] > array[i+1])
+			if (array[i] > array[i + 1])
 			{
 				temp = array[i];
-				array[i] = array[i+1];
-				array[i+1] = temp;
+				array[i] = array[i + 1];
+				array[i + 1] = temp;
 				print_array(array, size);
 			}
 		}
